:sparkles: Making SCHEDULE env optional

+++ backups/backup.sh
@@ -41,12 +41,14 @@ fi
 
 aws s3 cp "$LOCAL_FILENAME" "$S3_UPLOAD_URI"
 
-log "Removing old backups from S3..."
-S3_RETENTION_QUERY="Contents[?LastModified<='$(subtract-date-from-now $BACKUP_RETENTION_DAYS) 00:00:00'].{Key: Key}"
-aws s3api list-objects \
-	--bucket "${S3_BUCKET}" \
-	--prefix "${S3_PREFIX:-}" \
-	--query "${S3_RETENTION_QUERY}" \
-	--output text | xargs -P "$(nproc)" -I '{}' aws s3 rm "s3://${S3_BUCKET}/{}"
+if [ -n "$BACKUP_RETENTION_DAYS"]; then
+	log "Removing old backups from S3..."
+	S3_RETENTION_QUERY="Contents[?LastModified<='$(subtract-date-from-now $BACKUP_RETENTION_DAYS) 00:00:00'].{Key: Key}"
+	aws s3api list-objects \
+		--bucket "${S3_BUCKET}" \
+		--prefix "${S3_PREFIX:-}" \
+		--query "${S3_RETENTION_QUERY}" \
+		--output text | xargs -P "$(nproc)" -I '{}' aws s3 rm "s3://${S3_BUCKET}/{}"
+fi
 
 log "Backup successfully done"
